requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4
python-dotenv==1.0.1
pydantic[email]==2.6.4
python-multipart==0.0.9
